pyproject.toml
[tool.poetry]
name = "questions-app"
version = "0.1.0"
description = "Questions App"
authors = ["kioharu"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
alembic = "1.14.1"
fastapi = { extras = ["standard"], version = "0.115.7" }
gunicorn = { extras = ["gevent"], version = "23.0.0" }
redis = { extras = ["hiredis"], version = "6.1.0" }
pydantic-settings = "2.5.2"
sqlalchemy = "2.0.37"
asyncpg = "0.28.0"
python-jose = "3.4.0"
passlib = "1.7.4"
python-multipart = "0.0.20"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"
pre-commit = "*"
pyright = { extras = ["nodejs"], version = "*" }
ruff = "*"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "*"
pytest-asyncio = "*"
pytest-cov = "*"
factory-boy = "*"
faker = "*"
aiosqlite = "*"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
atomic = true
profile = "black"
py_version = 311
line_length = 88
skip_gitignore = true
multi_line_output = 3

[tool.ruff.lint]
select = [
    "F",    # Pyflakes
    "E",    # pycodestyle
    "W",    # pycodestyle
    "C90",  # mccabe
    "I",    # isort
    "N",    # pep8-naming
    "S",    # flake8-bandit
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "EM",   # flake8-errmsg
    "ISC",  # flake8-implicit-str-concat
    "ICN",  # flake8-import-conventions
    "LOG",  # flake8-logging
    "G",    # flake8-logging-format
    "PIE",  # flake8-pie
    "T20",  # flake8-print
    "PT",   # flake8-pytest-style
    "RSE",  # flake8-raise
    "RET",  # flake8-return
    "SIM",  # flake8-simplify
    "TC",   # flake8-type-checking
    "PTH",  # flake8-use-pathlib
    "PGH",  # pygrep-hooks
    "PL",   # Pylint
    "TRY",  # tryceratops
    "PERF", # Perflint
    "FURB", # refurb
    "FAST", # FastAPI
]
exclude = ["db/alembic/versions/*.py", "db/alembic/env.py"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = [
    "S101",    # "Use of `assert` detected"
    "PLR0913", # "Too many arguments in function definition"
    "S107",    # "Possible hardcoded password assigned to function default"
    "PGH003",  # "Use specific rule codes when ignoring type issues"
]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.11"
ignore = ["db"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::sqlalchemy.exc.MovedIn20Warning",
    "ignore::pydantic._internal._config.PydanticDeprecatedSince20"
]

[tool.coverage.run]
omit = ["tests/*", "db/alembic/versions/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
